refactoring: common and object_common to their own files
